# Fix `Image.convert()` overwriting custom mipmaps